[BUGFIX] Missing check on non-null intermediate airport
One of the requirements of the task was to ensure that the connecting
airport was null to have a route valid (apart from the operator check).
Imitating what was done in the connection validation, a decorator
pattern has been put in place to inject the validators of the routes and
so decoupling different parts of the business logic. Tests added as well
and modified existing.